backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
openpyxl
